Rework entire buildsystem and bootloader
This entire project can now be compiled using only the
Rust toolchain!

This rework replaces cargo-xbuild with the nightly build-std
feature.

The bootloader has been entirely rewritten to be used as
global_asm, this eliminates the need for NASM. However it now has
to be a two-stage bootloader as LLVM doesn't compact instructions
as well as NASM.

A custom runner as also been created to replace the need for UNIX
tools such as dd and Makefile, it also removes the need for
cargo-binutils.